+++ src/Payetools.Common/Model/UkPostcode.cs
@@ -13,18 +13,8 @@ namespace Payetools.Common.Model;
 /// </summary>
 public partial class UkPostcode
 {
-#if NET7_0_OR_GREATER
-
     [GeneratedRegex(@"^(([A-Z]{1,2}\d[A-Z\d]?|ASCN|STHL|TDCU|BBND|[BFS]IQQ|PCRN|TKCA) ?\d[A-Z]{2}|BFPO ?\d{1,4}|(KY\d|MSR|VG|AI)[ -]?\d{4}|[A-Z]{2} ?\d{2}|GE ?CX)$")]
     private static partial Regex GetValidationRegex();
-
-#else
-
-    private static readonly Regex _validationRegex = new Regex(@"^(([A-Z]{1,2}\d[A-Z\d]?|ASCN|STHL|TDCU|BBND|[BFS]IQQ|PCRN|TKCA) ?\d[A-Z]{2}|BFPO ?\d{1,4}|(KY\d|MSR|VG|AI)[ -]?\d{4}|[A-Z]{2} ?\d{2}|GE ?CX)$");
-
-    private static Regex GetValidationRegex() => _validationRegex;
-
-#endif
 
     private readonly string _value;