+++ test/Payetools.Common.Tests/DateOnlyExtensionTests.cs
@@ -4,7 +4,7 @@
 //
 //   * The MIT License, see https://opensource.org/license/mit/
 
-using FluentAssertions;
+using Shouldly;
 using Payetools.Common.Extensions;
 using System;
 using System.Collections.Generic;
@@ -21,26 +21,26 @@ public class DateOnlyExtensionTests
     {
         var date = new DateOnly(2022, 1, 1);
         var dateOfBirth = new DateOnly(2000, 1, 1);
-        dateOfBirth.AgeAt(date).Should().Be(22, $"date of birth = {dateOfBirth} and date = {date}");
+        dateOfBirth.AgeAt(date).ShouldBe(22, $"date of birth = {dateOfBirth} and date = {date}");
 
         date = new DateOnly(2022, 12, 31);
         dateOfBirth = new DateOnly(2000, 1, 1);
-        dateOfBirth.AgeAt(date).Should().Be(22, $"date of birth = {dateOfBirth} and date = {date}");
+        dateOfBirth.AgeAt(date).ShouldBe(22, $"date of birth = {dateOfBirth} and date = {date}");
 
         date = new DateOnly(2023, 1, 1);
         dateOfBirth = new DateOnly(2000, 1, 1);
-        dateOfBirth.AgeAt(date).Should().Be(23, $"date of birth = {dateOfBirth} and date = {date}");
+        dateOfBirth.AgeAt(date).ShouldBe(23, $"date of birth = {dateOfBirth} and date = {date}");
 
         date = new DateOnly(2023, 2, 28);
         dateOfBirth = new DateOnly(2000, 2, 28);
-        dateOfBirth.AgeAt(date).Should().Be(23, $"date of birth = {dateOfBirth} and date = {date}");
+        dateOfBirth.AgeAt(date).ShouldBe(23, $"date of birth = {dateOfBirth} and date = {date}");
 
         date = new DateOnly(2023, 2, 28);
         dateOfBirth = new DateOnly(2000, 2, 29);
-        dateOfBirth.AgeAt(date).Should().Be(22, $"date of birth = {dateOfBirth} and date = {date}");
+        dateOfBirth.AgeAt(date).ShouldBe(22, $"date of birth = {dateOfBirth} and date = {date}");
 
         date = new DateOnly(2023, 3, 1);
         dateOfBirth = new DateOnly(2000, 2, 29);
-        dateOfBirth.AgeAt(date).Should().Be(23, $"date of birth = {dateOfBirth} and date = {date}");
+        dateOfBirth.AgeAt(date).ShouldBe(23, $"date of birth = {dateOfBirth} and date = {date}");
     }
 }