+++ test/Payetools.Common.Tests/HmrcPayeReferenceTests.cs
@@ -4,7 +4,7 @@
 //
 //   * The MIT License, see https://opensource.org/license/mit/
 
-using FluentAssertions;
+using Shouldly;
 using Payetools.Common.Model;
 
 namespace Payetools.Common.Tests;
@@ -15,48 +15,48 @@ public class HmrcPayeReferenceTests
     public void TestValidPayeReferences()
     {
         var success = HmrcPayeReference.TryParse("123/ABC", out var result);
-        success.Should().BeTrue();
-        result.ToString().Should().Be("123/ABC");
-        result?.HmrcOfficeNumber.Should().Be(123);
-        result?.EmployerPayeReference.Should().Be("ABC");
+        success.ShouldBeTrue();
+        result.ToString().ShouldBe("123/ABC");
+        result?.HmrcOfficeNumber.ShouldBe(123);
+        result?.EmployerPayeReference.ShouldBe("ABC");
 
         success = HmrcPayeReference.TryParse(" 456 / ABC12345", out result);
-        success.Should().BeTrue();
-        result.ToString().Should().Be("456/ABC12345");
-        result?.HmrcOfficeNumber.Should().Be(456);
-        result?.EmployerPayeReference.Should().Be("ABC12345");
+        success.ShouldBeTrue();
+        result.ToString().ShouldBe("456/ABC12345");
+        result?.HmrcOfficeNumber.ShouldBe(456);
+        result?.EmployerPayeReference.ShouldBe("ABC12345");
     }
 
     [Fact]
     public void TestInvalidAccountsOfficeReferences()
     {
         var success = HmrcPayeReference.TryParse("1/ABC", out var result);
-        success.Should().BeFalse();
-        result.Should().BeNull();
+        success.ShouldBeFalse();
+        result.ShouldBeNull();
 
         success = HmrcPayeReference.TryParse("123/ABC1212121212121", out result);
-        success.Should().BeFalse();
-        result.Should().BeNull();
+        success.ShouldBeFalse();
+        result.ShouldBeNull();
 
         success = HmrcPayeReference.TryParse("123/?$%", out result);
-        success.Should().BeFalse();
-        result.Should().BeNull();
+        success.ShouldBeFalse();
+        result.ShouldBeNull();
 
         success = HmrcPayeReference.TryParse(null, out result);
-        success.Should().BeFalse();
-        result.Should().BeNull();
+        success.ShouldBeFalse();
+        result.ShouldBeNull();
 
         success = HmrcPayeReference.TryParse(string.Empty, out result);
-        success.Should().BeFalse();
-        result.Should().BeNull();
+        success.ShouldBeFalse();
+        result.ShouldBeNull();
     }
 
     [Fact]
     public void TestImplicitCast()
     {
         var success = HmrcPayeReference.TryParse("123 / ABC", out var result);
-        success.Should().BeTrue();
+        success.ShouldBeTrue();
         string stringResult = result ?? "";
-        stringResult.Should().Be("123/ABC");
+        stringResult.ShouldBe("123/ABC");
     }
 }