+++ test/Payetools.Common.Tests/NiNumberTests.cs
@@ -4,7 +4,7 @@
 //
 //   * The MIT License, see https://opensource.org/license/mit/
 
-using FluentAssertions;
+using Shouldly;
 using Payetools.Common.Model;
 
 namespace Payetools.Common.Tests;
@@ -15,19 +15,19 @@ public class NiNumberTests
     public void TestValidNiNumbers()
     {
         var niNumber = new NiNumber("AB123456A");
-        niNumber.ToString().Should().Be("AB123456A");
+        niNumber.ToString().ShouldBe("AB123456A");
 
         niNumber = new NiNumber("HJ 12 34 56 B");
-        niNumber.ToString().Should().Be("HJ123456B");
+        niNumber.ToString().ShouldBe("HJ123456B");
 
         niNumber = new NiNumber("KL 12 34 56 C");
-        niNumber.ToString().Should().Be("KL123456C");
+        niNumber.ToString().ShouldBe("KL123456C");
 
         niNumber = new NiNumber("  PR12 3456 D  ");
-        niNumber.ToString().Should().Be("PR123456D");
+        niNumber.ToString().ShouldBe("PR123456D");
 
         niNumber = new NiNumber("ab123456a");
-        niNumber.ToString().Should().Be("AB123456A");
+        niNumber.ToString().ShouldBe("AB123456A");
     }
 
     [Fact]
@@ -61,28 +61,28 @@ public class NiNumberTests
     public void TestImplicitCasts()
     {
         string value = new NiNumber("bc654321d");
-        value.Should().Be("BC654321D");
+        value.ShouldBe("BC654321D");
 
         NiNumber niNumber = "EG 12 45  78 A";
-        niNumber.ToString().Should().Be("EG124578A");
+        niNumber.ToString().ShouldBe("EG124578A");
     }
 
     [Fact]
     public void TestToStringMethods()
     {
         var value = new NiNumber("bc654321d");
-        value.ToString().Should().Be("BC654321D");
-        value.ToString(true).Should().Be("BC 65 43 21 D");
+        value.ToString().ShouldBe("BC654321D");
+        value.ToString(true).ShouldBe("BC 65 43 21 D");
     }
 
     private void TestInvalidNiNumber(string niNumber)
     {
         Action action = () => new NiNumber(niNumber);
-        action.Should().Throw<ArgumentException>()
-            .WithMessage("Argument is not a valid National Insurance Number (Parameter 'niNumber')");
+        action.ShouldThrow<ArgumentException>()
+            .Message.ShouldBe("Argument is not a valid National Insurance Number (Parameter 'niNumber')");
 
         action = () => new NiNumber(niNumber.ToLower());
-        action.Should().Throw<ArgumentException>()
-            .WithMessage("Argument is not a valid National Insurance Number (Parameter 'niNumber')");
+        action.ShouldThrow<ArgumentException>()
+            .Message.ShouldBe("Argument is not a valid National Insurance Number (Parameter 'niNumber')");
     }
 }