+++ test/Payetools.Common.Tests/TaxCodeTestHelper.cs
@@ -4,7 +4,7 @@
 //
 //   * The MIT License, see https://opensource.org/license/mit/
 
-using FluentAssertions;
+using Shouldly;
 using Payetools.Common.Model;
 
 namespace Payetools.Common.Tests;
@@ -17,27 +17,27 @@ public static class TaxCodeTestHelper
     {
         var result = TaxCode.TryParse(input.ToLower(), _testTaxYear, out var taxCode);
 
-        result.Should().BeTrue();
+        result.ShouldBeTrue();
         result.ShouldNotHaveDefaultValue();
-        taxCode.TaxTreatment.Should().Be(expectedTreatment);
+        taxCode.TaxTreatment.ShouldBe(expectedTreatment);
 
         result = TaxCode.TryParse(input.ToUpper(), out taxCode);
 
-        result.Should().BeTrue();
+        result.ShouldBeTrue();
         result.ShouldNotHaveDefaultValue();
-        taxCode.TaxTreatment.Should().Be(expectedTreatment);
+        taxCode.TaxTreatment.ShouldBe(expectedTreatment);
     }
 
     public static void RunInvalidCodeTest(string input)
     {
         var result = TaxCode.TryParse(input.ToLower(), _testTaxYear, out var taxCode);
 
-        result.Should().BeFalse();
+        result.ShouldBeFalse();
         result.ShouldHaveDefaultValue();
 
         result = TaxCode.TryParse(input.ToUpper(), out taxCode);
 
-        result.Should().BeFalse();
+        result.ShouldBeFalse();
         result.ShouldHaveDefaultValue();
     }
 
@@ -45,16 +45,16 @@ public static class TaxCodeTestHelper
     {
         var result = TaxCode.TryParse(input.ToLower(), _testTaxYear, out var taxCode);
 
-        result.Should().BeTrue();
+        result.ShouldBeTrue();
         result.ShouldNotHaveDefaultValue();
-        taxCode.IsNonCumulative.Should().BeTrue();
+        taxCode.IsNonCumulative.ShouldBeTrue();
 
         result = TaxCode.TryParse(input.ToUpper(), out taxCode);
 
-        result.Should().BeTrue();
+        result.ShouldBeTrue();
         result.ShouldNotHaveDefaultValue();
-        taxCode.IsNonCumulative.Should().BeTrue();
-        taxCode.TaxTreatment.Should().Be(expectedTreatment);
+        taxCode.IsNonCumulative.ShouldBeTrue();
+        taxCode.TaxTreatment.ShouldBe(expectedTreatment);
     }
 
     public static void RunFixedCodeCountrySpecificTest(string input,
@@ -64,17 +64,17 @@ public static class TaxCodeTestHelper
     {
         var result = TaxCode.TryParse(input.ToLower(), taxYear, out var taxCode);
 
-        result.Should().BeTrue();
+        result.ShouldBeTrue();
         result.ShouldNotHaveDefaultValue();
-        taxCode.TaxTreatment.Should().Be(expectedTreatment);
-        taxCode.ApplicableCountries.Should().Be(expectedCountries);
+        taxCode.TaxTreatment.ShouldBe(expectedTreatment);
+        taxCode.ApplicableCountries.ShouldBe(expectedCountries);
 
         result = TaxCode.TryParse(input.ToUpper(), taxYear, out taxCode);
 
-        result.Should().BeTrue();
+        result.ShouldBeTrue();
         result.ShouldNotHaveDefaultValue();
-        taxCode.TaxTreatment.Should().Be(expectedTreatment);
-        taxCode.ApplicableCountries.Should().Be(expectedCountries);
+        taxCode.TaxTreatment.ShouldBe(expectedTreatment);
+        taxCode.ApplicableCountries.ShouldBe(expectedCountries);
     }
 
     public static void RunStandardCodeTest(string input,
@@ -85,40 +85,40 @@ public static class TaxCodeTestHelper
     {
         var result = TaxCode.TryParse(input.ToLower(), taxYear, out var taxCode);
 
-        result.Should().BeTrue();
+        result.ShouldBeTrue();
         result.ShouldNotHaveDefaultValue();
-        taxCode.TaxTreatment.Should().Be(expectedTreatment);
-        taxCode.ApplicableCountries.Should().Be(expectedCountries);
-        taxCode.NotionalAllowance.Should().Be(expectedAllowance);
+        taxCode.TaxTreatment.ShouldBe(expectedTreatment);
+        taxCode.ApplicableCountries.ShouldBe(expectedCountries);
+        taxCode.NotionalAllowance.ShouldBe(expectedAllowance);
 
         result = TaxCode.TryParse(input.ToUpper(), taxYear, out taxCode);
 
-        result.Should().BeTrue();
+        result.ShouldBeTrue();
         result.ShouldNotHaveDefaultValue();
-        taxCode.TaxTreatment.Should().Be(expectedTreatment);
-        taxCode.ApplicableCountries.Should().Be(expectedCountries);
-        taxCode.NotionalAllowance.Should().Be(expectedAllowance);
+        taxCode.TaxTreatment.ShouldBe(expectedTreatment);
+        taxCode.ApplicableCountries.ShouldBe(expectedCountries);
+        taxCode.NotionalAllowance.ShouldBe(expectedAllowance);
     }
 
     public static void RunToStringTest(string input, string expectedOutput, bool expectedIsNonCumulative)
     {
-        TaxCode.TryParse(input.ToLower(), _testTaxYear, out var taxCode).Should().BeTrue();
-        taxCode.ToString().Should().Be(expectedOutput);
-        taxCode.IsNonCumulative.Should().Be(expectedIsNonCumulative);
+        TaxCode.TryParse(input.ToLower(), _testTaxYear, out var taxCode).ShouldBeTrue();
+        taxCode.ToString().ShouldBe(expectedOutput);
+        taxCode.IsNonCumulative.ShouldBe(expectedIsNonCumulative);
 
-        TaxCode.TryParse(input.ToUpper(), _testTaxYear, out taxCode).Should().BeTrue();
-        taxCode.ToString().Should().Be(expectedOutput);
-        taxCode.IsNonCumulative.Should().Be(expectedIsNonCumulative);
+        TaxCode.TryParse(input.ToUpper(), _testTaxYear, out taxCode).ShouldBeTrue();
+        taxCode.ToString().ShouldBe(expectedOutput);
+        taxCode.IsNonCumulative.ShouldBe(expectedIsNonCumulative);
     }
 
     public static void RunToFullStringTest(string input, string expectedOutput, bool expectedIsNonCumulative)
     {
-        TaxCode.TryParse(input.ToLower(), _testTaxYear, out var taxCode).Should().BeTrue();
-        taxCode.ToString(true, true).Should().Be(expectedOutput);
-        taxCode.IsNonCumulative.Should().Be(expectedIsNonCumulative);
+        TaxCode.TryParse(input.ToLower(), _testTaxYear, out var taxCode).ShouldBeTrue();
+        taxCode.ToString(true, true).ShouldBe(expectedOutput);
+        taxCode.IsNonCumulative.ShouldBe(expectedIsNonCumulative);
 
-        TaxCode.TryParse(input.ToUpper(), _testTaxYear, out taxCode).Should().BeTrue();
-        taxCode.ToString(true, true).Should().Be(expectedOutput);
-        taxCode.IsNonCumulative.Should().Be(expectedIsNonCumulative);
+        TaxCode.TryParse(input.ToUpper(), _testTaxYear, out taxCode).ShouldBeTrue();
+        taxCode.ToString(true, true).ShouldBe(expectedOutput);
+        taxCode.IsNonCumulative.ShouldBe(expectedIsNonCumulative);
     }
 }