+++ test/Payetools.Common.Tests/TaxYearTests.cs
@@ -4,7 +4,7 @@
 //
 //   * The MIT License, see https://opensource.org/license/mit/
 
-using FluentAssertions;
+using Shouldly;
 using Payetools.Common.Model;
 
 namespace Payetools.Common.Tests;
@@ -16,15 +16,15 @@ public class TaxYearTests
     {
         var date = new DateOnly(2020, 5, 5);
         var taxYear = new TaxYear(date);
-        taxYear.TaxYearEnding.Should().Be(TaxYearEnding.Apr5_2021);
+        taxYear.TaxYearEnding.ShouldBe(TaxYearEnding.Apr5_2021);
 
         date = new DateOnly(2022, 4, 5);
         taxYear = new TaxYear(date);
-        taxYear.TaxYearEnding.Should().Be(TaxYearEnding.Apr5_2022);
+        taxYear.TaxYearEnding.ShouldBe(TaxYearEnding.Apr5_2022);
 
         date = new DateOnly(2022, 4, 6);
         taxYear = new TaxYear(date);
-        taxYear.TaxYearEnding.Should().Be(TaxYearEnding.Apr5_2023);
+        taxYear.TaxYearEnding.ShouldBe(TaxYearEnding.Apr5_2023);
     }
 
     [Fact]
@@ -117,8 +117,8 @@ public class TaxYearTests
             var periodNumber = taxYear.GetTaxPeriod(new DateOnly(2022, 4, 6), PayFrequency.Monthly);
         };
 
-        action.Should().Throw<ArgumentException>()
-            .WithMessage("Pay date of 06/04/2022 is outside this tax year 06/04/2021 - 05/04/2022 (Parameter 'payDate')");
+        action.ShouldThrow<ArgumentException>()
+            .Message.ShouldBe("Pay date of 06/04/2022 is outside this tax year 06/04/2021 - 05/04/2022 (Parameter 'payDate')");
     }
 
     [Fact]
@@ -126,18 +126,18 @@ public class TaxYearTests
     {
         var taxYear = new TaxYear(TaxYearEnding.Apr5_2022);
 
-        taxYear.GetLastDayOfTaxPeriod(PayFrequency.Weekly, 1).Should().Be(new DateOnly(2021, 4, 12));
-        taxYear.GetLastDayOfTaxPeriod(PayFrequency.Weekly, 2).Should().Be(new DateOnly(2021, 4, 19));
-        taxYear.GetLastDayOfTaxPeriod(PayFrequency.Fortnightly, 1).Should().Be(new DateOnly(2021, 4, 19));
-        taxYear.GetLastDayOfTaxPeriod(PayFrequency.Fortnightly, 2).Should().Be(new DateOnly(2021, 5, 3));
-        taxYear.GetLastDayOfTaxPeriod(PayFrequency.FourWeekly, 1).Should().Be(new DateOnly(2021, 5, 3));
-        taxYear.GetLastDayOfTaxPeriod(PayFrequency.Monthly, 1).Should().Be(new DateOnly(2021, 5, 5));
-        taxYear.GetLastDayOfTaxPeriod(PayFrequency.Monthly, 2).Should().Be(new DateOnly(2021, 6, 5));
-        taxYear.GetLastDayOfTaxPeriod(PayFrequency.Monthly, 3).Should().Be(new DateOnly(2021, 7, 5));
-        taxYear.GetLastDayOfTaxPeriod(PayFrequency.Monthly, 12).Should().Be(new DateOnly(2022, 4, 5));
-        taxYear.GetLastDayOfTaxPeriod(PayFrequency.Quarterly, 1).Should().Be(new DateOnly(2021, 7, 5));
-        taxYear.GetLastDayOfTaxPeriod(PayFrequency.BiAnnually, 1).Should().Be(new DateOnly(2021, 10, 5));
-        taxYear.GetLastDayOfTaxPeriod(PayFrequency.Annually, 1).Should().Be(new DateOnly(2022, 4, 5));
+        taxYear.GetLastDayOfTaxPeriod(PayFrequency.Weekly, 1).ShouldBe(new DateOnly(2021, 4, 12));
+        taxYear.GetLastDayOfTaxPeriod(PayFrequency.Weekly, 2).ShouldBe(new DateOnly(2021, 4, 19));
+        taxYear.GetLastDayOfTaxPeriod(PayFrequency.Fortnightly, 1).ShouldBe(new DateOnly(2021, 4, 19));
+        taxYear.GetLastDayOfTaxPeriod(PayFrequency.Fortnightly, 2).ShouldBe(new DateOnly(2021, 5, 3));
+        taxYear.GetLastDayOfTaxPeriod(PayFrequency.FourWeekly, 1).ShouldBe(new DateOnly(2021, 5, 3));
+        taxYear.GetLastDayOfTaxPeriod(PayFrequency.Monthly, 1).ShouldBe(new DateOnly(2021, 5, 5));
+        taxYear.GetLastDayOfTaxPeriod(PayFrequency.Monthly, 2).ShouldBe(new DateOnly(2021, 6, 5));
+        taxYear.GetLastDayOfTaxPeriod(PayFrequency.Monthly, 3).ShouldBe(new DateOnly(2021, 7, 5));
+        taxYear.GetLastDayOfTaxPeriod(PayFrequency.Monthly, 12).ShouldBe(new DateOnly(2022, 4, 5));
+        taxYear.GetLastDayOfTaxPeriod(PayFrequency.Quarterly, 1).ShouldBe(new DateOnly(2021, 7, 5));
+        taxYear.GetLastDayOfTaxPeriod(PayFrequency.BiAnnually, 1).ShouldBe(new DateOnly(2021, 10, 5));
+        taxYear.GetLastDayOfTaxPeriod(PayFrequency.Annually, 1).ShouldBe(new DateOnly(2022, 4, 5));
     }
 
     private static void RunTaxPeriodTest(
@@ -151,19 +151,19 @@ public class TaxYearTests
         var taxYear = new TaxYear(taxYearEnding);
         var periodNumber = taxYear.GetTaxPeriod(payDate, payFrequency);
 
-        periodNumber.Should().Be(expectedPeriodNumber);
+        periodNumber.ShouldBe(expectedPeriodNumber);
 
         if (expectedWeekNumber != -1)
         {
             var weekNumber = taxYear.GetWeekNumber(payDate, payFrequency);
 
             var expWkNo = expectedWeekNumber != 0 ? expectedWeekNumber : expectedPeriodNumber;
-            weekNumber.Should().Be(expWkNo);
+            weekNumber.ShouldBe(expWkNo);
         }
 
         var monthNumber = taxYear.GetMonthNumber(payDate, payFrequency);
 
         var expMnthNo = expectedMonthNumber != 0 ? expectedMonthNumber : expectedPeriodNumber;
-        monthNumber.Should().Be(expMnthNo);
+        monthNumber.ShouldBe(expMnthNo);
     }
 }